Add files via upload
This is a Report of Banking Loan System which is made with the composition of Power BI and SQL using MY SQL server. This report helps in Analysing and Simplifying various aspects like average Interest Rate , average DTI, Total funded amounts and Total Applications. It also distinguishes Good Loan and Bad Loan which helps the user to work better and have a good knowledge. It contains various charts to show trends in Issue Date of Loan and Loan Purpose Break Down as well as Tree Maps for Homeownership Analysis in the Overview dashboard . It also contains a GRID which provides a consolidated view of essential information within the Loan Data and provides holistic snapshots of Loan related metrices and data points.